// src/app/components/forms/GuestForm/constants/enums.ts
export enum FormStep {
  GUEST_INFO = 0,
  THEME_PREFERENCES = 1,
  CONFIRMATION = 2
}

export enum ThemePreference {
  STAR_WARS = 'star_wars',
  HARRY_POTTER = 'harry_potter',
  BOTH = 'both',
  NONE = 'none'
}

export enum HogwartsHouse {
  GRYFFINDOR = 'gryffindor',
  SLYTHERIN = 'slytherin',
  HUFFLEPUFF = 'hufflepuff',
  RAVENCLAW = 'ravenclaw'
}

export enum StarWarsSide {
  JEDI = 'jedi',
  SITH = 'sith'
} 